+++ src/keys.rs
@@ -38,8 +38,18 @@ pub fn generate_seed(
       full_entropy = wallet.seed_secret.full_entropy.clone();
       mnemonic_dictionary = wallet.seed_secret.mnemonic_dictionary.clone();
     }
+    "QRNG" => {
+      let raw_entropy: Zeroizing<String> = wallet.seed_secret.raw_entropy.clone();
+
+      let entropy_checksum: Zeroizing<String> = e_q::calculate_checksum_for_entropy(raw_entropy.clone());
+      wallet.seed_secret.entropy_checksum = entropy_checksum.clone();
+
+      full_entropy = Zeroizing::new(format!("{}{}", *raw_entropy, *entropy_checksum));
+      wallet.seed_secret.full_entropy = full_entropy.clone();
+
+      mnemonic_dictionary = wallet.seed_secret.mnemonic_dictionary.clone();
+    }
     "RNG" => {
-      // TODO: Implement RNG Mnemonic passphrase + add to GUI
       let entropy_length: Zeroizing<usize> = wallet.seed_secret.entropy_length.clone();
       mnemonic_dictionary = wallet.seed_secret.mnemonic_dictionary.clone();
 
@@ -51,21 +61,23 @@ pub fn generate_seed(
 
       full_entropy = Zeroizing::new(format!("{}{}", *raw_entropy, *entropy_checksum));
       wallet.seed_secret.full_entropy = full_entropy.clone();
-
-      match wallet.seed_secret.mnemonic_passphrase_source.as_str() {
-        "RNG" => {
-          match generate_raw_mnemonic_passphrase(MNEMONIC_PASSPHRASE_LENGTH as usize) {
-            Ok(pass) => {
-              wallet.seed_secret.mnemonic_passphrase = pass;
-            }
-            Err(err) => return Err(AppError::log(format!("Error: {:?}", err))),
-          };
-        }
-        "Custom" => {}
-        _ => {}
-      }
     }
-    _ => return Err(AppError::log(format!("Unknown entropy source: {:?}", entropy_source))),
+    _ => {
+      return Err(AppError::log(format!("Unknown entropy source: {:?}", entropy_source)));
+    }
+  }
+
+  match wallet.seed_secret.mnemonic_passphrase_source.as_str() {
+    "RNG" => {
+      match generate_raw_mnemonic_passphrase(MNEMONIC_PASSPHRASE_LENGTH as usize) {
+        Ok(pass) => {
+          wallet.seed_secret.mnemonic_passphrase = pass;
+        }
+        Err(err) => return Err(AppError::log(format!("Error: {:?}", err))),
+      };
+    }
+    "Custom" => {}
+    _ => {}
   }
 
   let mnemonic_words: Zeroizing<String> = match generate_mnemonic_words(full_entropy.clone(), Some(mnemonic_dictionary)) {